+++ lizard_fewsjdbc/restapi/__init__.py
@@ -1,0 +1,1 @@
+# (c) Copyright Nelen & Schuurmans 2012