+++ lizard_fewsjdbc/restapi/renderers.py
@@ -1,0 +1,4 @@
+from djangorestframework.renderers import DocumentingHTMLRenderer
+from djangorestframework.renderers import JSONRenderer
+
+RENDERERS = (JSONRenderer, DocumentingHTMLRenderer,)